+++ cypress-shared/support/fedex-shipping/sla.js
@@ -1,0 +1,6 @@
+export default {
+  FedexHomeDelivery: 'FedEx Home Delivery',
+  FirstOvernight: 'First Overnight',
+  FedexGroundDelivery: 'FedEx Ground',
+  InternationalEconomy: 'International Economy',
+}